+++ src/components/ProductSlide/ProductCard/ProductCard.tsx
@@ -1,4 +1,4 @@
-import { Image, View } from "react-native";
+import { Image, TouchableOpacity, View } from "react-native";
 import { Typography } from "../../Typography";
 import { Product } from "../ProductSlide";
 import { ProductCardHOC } from "../../../hoc/ProductsCardStyle";
@@ -8,27 +8,34 @@ interface Props {
   style?: any
 }
 
-const ProductCard: React.FC<Props> = ({ product, style }) => {
+const ProductCardComponent: React.FC<Props> = ({ product, style }) => {
+
+  const handleButtonClick = () => {
+    console.log(product.title)
+  }
+
   return (
-    <View style={[style.productCardContainer, style]} key={product.id}>
-      <Image
-        source={{
-          uri: product.images[0],
-        }}
-        style={style.thumbnail}
-      />
-      <Typography type="largeParagraph" style={style.productTitle}>
-        {product.title}
-      </Typography>
-      <Typography type="paragraph" style={style.productDescription}>
-        {product.description}
-      </Typography>
-      <Typography type="paragraph" style={style.productPrice}>
-        R$ {product.price}
-      </Typography>
-    </View>
+    <TouchableOpacity onPress={() => handleButtonClick()}>
+      <View style={[style.productCardContainer, style]} key={product.id}>
+        <Image
+          source={{
+            uri: product.images[0],
+          }}
+          style={style.thumbnail}
+        />
+        <Typography type="largeParagraph" style={style.productTitle}>
+          {product.title}
+        </Typography>
+        <Typography type="paragraph" style={style.productDescription}>
+          {product.description}
+        </Typography>
+        <Typography type="paragraph" style={style.productPrice}>
+          R$ {product.price}
+        </Typography>
+      </View>
+    </TouchableOpacity>
   );
 };
 
-export const ProductCardTest = ProductCardHOC(ProductCard);
+export const ProductCard = ProductCardHOC(ProductCardComponent);
