+++ src/components/SearchProducts/SearchProductList/SearchProductList.tsx
@@ -2,8 +2,8 @@ import { FlatList, Image, View } from "react-native";
 import { Typography } from "../../Typography";
 import { styles } from "./styles";
 import { Product } from "../../ProductSlide";
-import SearchProducts from "../SearchProducts";
 import { useEffect } from "react";
+import { ProductCard } from "../../ProductSlide/ProductCard";
 
 interface Props {
   data: Product[],
@@ -11,14 +11,14 @@ interface Props {
 
 export const SearchProductList: React.FC<Props> = ({data}) => {
   return (
-    <View style={{flex: 1}}>
+    <View style={styles.productListContainer}>
       <Typography type="paragraph" style={styles.productFoundText}> {data.length} products found </Typography>
       <View style={styles.productListContainer} >
         <FlatList
           data={data}
           numColumns={2}
           keyExtractor={(item) => item.id}
-          renderItem={({ item, index }) => <SearchProducts product={item} position={index} />}
+          renderItem={({ item, index }) => <ProductCard product={item} index={index} search />}
         />
       </View>
     </View>
