--- src/components/SearchProducts/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchProducts";
