--- src/components/SearchProducts/styles.ts
@@ -1,11 +1,0 @@
-import { StyleSheet } from "react-native";
-import { colorPalette } from "../../theme/colors";
-
-export const styles = StyleSheet.create({
-    productCardContainer: {
-        alignItems: "center",
-        padding: 10,
-        marginRight: 20,
-        shadowColor: colorPalette.dark,
-    }
-});
